+++ lib/paperplane_exceptions.dart
@@ -7,7 +7,9 @@
 /// It helps to manage different type of exception inside my library.
 library paperplane_exceptions;
 
+export 'package:paperplane/src/exceptions/dispatcher_exception.dart';
 export 'package:paperplane/src/exceptions/http_client_exception.dart';
 export 'package:paperplane/src/exceptions/paperplane_exception.dart';
 export 'package:paperplane/src/exceptions/telegram_api_exception.dart';
 export 'package:paperplane/src/exceptions/telegram_method_exception.dart';
+export 'package:paperplane/src/exceptions/updater_exception.dart';